+++ lib/src/platform_api.dart
@@ -1,0 +1,1 @@
+export 'native_api.dart' if (dart.library.js_interop) 'web_api.dart';